pyproject.toml
[project]
name = "auto-link-tyut-school-net"
version = "0.1.0"
description = "TYUT自动登录工具"
authors = [{name = "3113y"}]
dependencies = [
    "pyqt5==5.15.11",
    "onnxruntime==1.19.2",
    "pillow==11.3.0",
    "altgraph==0.17.4",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "coloredlogs==15.0.1",
    "flatbuffers==25.9.23",
    "humanfriendly==10.0",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "mpmath==1.3.0",
    "numpy==2.0.2",
    "opencv-python-headless==4.12.0.88",
    "packaging==25.0",
    "pefile==2023.2.7",
    "pip==22.0.4",
    "protobuf==6.33.0",
    "pyinstaller==6.16.0",
    "pyinstaller-hooks-contrib==2025.9",
    "pyqt5-qt5==5.15.2",
    "pyqt5-sip==12.17.1",
    "pyqtwebengine==5.15.7",
    "pyqtwebengine-qt5==5.15.2",
    "pyreadline3==3.5.4",
    "pywin32-ctypes==0.2.3",
    "requests==2.32.5",
    "setuptools==58.1.0",
    "sympy==1.14.0",
    "urllib3==2.5.0",
    "zipp==3.23.0",
    "img2pdf>=0.6.1",
    "pyyaml>=6.0.3",
    "jmcomic>=2.6.10",
]
requires-python = ">=3.10"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
